requirements.txt
# requirements.txt
python-binance==3.1.1
python-telegram-bot==20.3
pandas==2.0.3
finta==1.3
numpy==1.24.3
scikit-learn==1.3.0
google-generativeai==0.3.0
requests==2.31.0
lightgbm==4.1.0
joblib==1.3.2
schedule==1.2.0
python-dotenv==1.0.0